Исправления по замечаниям ревьюера
