+++ ImageFeed/ImageFeed/Assets.xcassets/Contents.json
@@ -1,6 +1,6 @@
 {
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
+    "info" : {
+        "author" : "xcode",
+        "version" : 1
+    }
 }
