+++ ImageFeed/ImageFeed/Assets.xcassets/like_button_off.imageset/Contents.json
@@ -1,11 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "like_button_off.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
       "filename" : "like_button_off@2x.png",
       "idiom" : "universal",
       "scale" : "2x"
