+++ ImageFeed/ImageFeed/Assets.xcassets/like_button_on.imageset/Contents.json
@@ -1,11 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "like_button_on.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
       "filename" : "like_button_on@2x.png",
       "idiom" : "universal",
       "scale" : "2x"
