+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/0.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "0.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "0 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
