+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/1.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "1.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "1 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
