+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/10.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "10.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "10 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
