+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/11.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "11.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "11 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
