+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/12.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "12.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "12 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
