+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/13.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "13.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "13 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
