+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/15.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "15.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "15 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
