+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/16.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "16.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "16 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
