+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/17.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "17.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "17 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
