+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/18.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "18.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "18 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
