+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/2.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "2.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "2 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
