+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/3.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "3.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "3 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
