+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/4.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "4.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "4 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
