+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/5.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "5.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "5 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
