+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/6.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "6.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "6 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
