+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/7.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "7.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "7 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
