+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/8.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "8.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "8 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
