+++ ImageFeed/ImageFeed/Assets.xcassets/mockImagesForTable/9.imageset/Contents.json
@@ -3,13 +3,10 @@
     {
       "filename" : "9.jpg",
       "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "9 1.jpg",
       "idiom" : "universal",
       "scale" : "3x"
     }
