+++ ImageFeed/ImageFeed/Assets.xcassets/splash_screen_logo.imageset/Contents.json
@@ -1,11 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "splash_screen_logo.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
       "filename" : "splash_screen_logo@2x.png",
       "idiom" : "universal",
       "scale" : "2x"
